Create ReadMe.MD
Predictive Models I
This module covers the basics of regression analysis. Our emphasis will be on applications and interpretation of the results for making real life business/policy decisions and not on the mathematical and statistical properties of the techniques used to produce these results. In order to provide a broad intuition of the concepts and methods, we will use data and examples from marketing decision making such as segmentation, estimating market potential and forecasting demand, etc.
Predictive Models II
We will continue covering regression analysis including non-linear transformations and dummy variables. In addition to advanced regression, we will cover prediction and classification trees.